refactor: convert CategoryBanner into general widget
