+++ packages/app_center/lib/src/games/games_page.dart
@@ -1,4 +1,3 @@
-import 'package:app_center/explore.dart';
 import 'package:app_center/games.dart';
 import 'package:app_center/l10n.dart';
 import 'package:app_center/layout.dart';
@@ -46,7 +45,6 @@ class GamesPage extends ConsumerWidget {
                       height: _CategoryBannerProperties.height,
                       kMaxSize: _CategoryBannerProperties.kMaxSize,
                       kIconSize: _CategoryBannerProperties.kIconSize,
-                      fontSize: _CategoryBannerProperties.fontSize,
                     )),
                 Container(width: 12),
                 const Expanded(
@@ -57,7 +55,6 @@ class GamesPage extends ConsumerWidget {
                       height: _CategoryBannerProperties.height,
                       kMaxSize: _CategoryBannerProperties.kMaxSize,
                       kIconSize: _CategoryBannerProperties.kIconSize,
-                      fontSize: _CategoryBannerProperties.fontSize,
                     ))
               ],
             ),
@@ -72,7 +69,6 @@ class GamesPage extends ConsumerWidget {
                       height: _CategoryBannerProperties.height,
                       kMaxSize: _CategoryBannerProperties.kMaxSize,
                       kIconSize: _CategoryBannerProperties.kIconSize,
-                      fontSize: _CategoryBannerProperties.fontSize,
                     )),
                 Container(width: 12),
                 const Expanded(
@@ -83,7 +79,6 @@ class GamesPage extends ConsumerWidget {
                       height: _CategoryBannerProperties.height,
                       kMaxSize: _CategoryBannerProperties.kMaxSize,
                       kIconSize: _CategoryBannerProperties.kIconSize,
-                      fontSize: _CategoryBannerProperties.fontSize,
                     ))
               ],
             ),
@@ -314,5 +309,4 @@ class _CategoryBannerProperties {
   static const double height = 150;
   static const double kMaxSize = 60;
   static const double kIconSize = 32;
-  static const double fontSize = 16;
 }
